all changes with update ui

+++ lib/model/Sales Person target model.dart
@@ -1,6 +1,6 @@
 class SalesPersonWiseTarget {
-  int? totalTarget;
-  double? totalAchieved;
+  double? totalTarget;
+  int? totalAchieved;
   double? totalVariance;
   String? salesPerson;
   String? itemGroup;
